app/vlm_server/pyproject.toml
[project]
name = "gnosis"
version = "0.1.0"
description = "VLM Inference Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "dotenv>=0.9.9",
    "google-genai>=1.46.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "huggingface-hub>=0.36.0",
    "modal>=1.2.1",
    "pre-commit>=4.4.0",
    "protobuf>=6.33.0",
    "pydantic>=2.12.3",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]

[tool.ruff]
line-length = 88
target-version = "py313"
extend-exclude = [
  "data",
  "build/",
  "dist/",
  ".venv/",
  "__pycache__",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
# which rule sets to enable; you can start simple
select = ["E", "F"]  # errors (E) + pyflakes (F)
ignore = []           # none to ignore initially
